\name{reviews}
\alias{reviews}
\docType{data}
\title{
Reviewers' scores
}
\description{
Reviewers' scores
}
\usage{data("reviews")}

\details{
TBD
}
\source{
The dataset is a part of The Riddle of Literary Quality Project.
}
\examples{
data(reviews)

# this section is not ready yet
# print(reviews)
# summary(reviews)

}
\keyword{datasets}
