\name{litRiddle-package}
\alias{litRiddle-package}
\alias{litRiddle}
\docType{package}
\title{
R Package to Research the Riddle of Literary Quality
}
\description{
The package contains the data of a reader survey about fiction in Dutch, a description of the novels the readers rated, and the results of stylistic measurements of the novels. The package also contains functions to combine, analyze, and visualize these data.
}
\details{
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Maciej Eder, Joris van Zundert, Corina Koolen, Karina van Dalen-Oskam
}
\references{
~~ Literature or other references for background information ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
